quick push on triglock for video
This has the temp frequency/amplitude blocks and a puzzle added to the left of the first unit circle puzzle. the LR/UD blocks are still the sin/cos asset for teleport. The bug is fixed. Also the current equation is displayed above the graph. It is red when wrong and green when correct.